Adds an optional BDD testing idiom.
The 'whiskey' package now exports a 'bdd' symbol which can be required to enable BDD-style tests:

var bdd = require('whiskey').bdd.init(exports);

This wraps the exports object of the test module and translates the BDD tests into whiskey tests
so they run normally.

The one adaptation of the BDD idiom for whiskey is that test suite and spec functions being passed
to describe() and it() need to accept a function parameter, respectively it() and expect() - this
is to support the whiskey behavior of injecting (test, assert) into each test individually.